// src/types/Anilist/RecentEP.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface RecentEP {
    currentPage:  number;
    hasNextPage:  boolean;
    totalPages:   number;
    totalResults: number;
    results:      Result[];
}

export interface Result {
    id:            string;
    malId:         number;
    title:         Title;
    image:         string;
    rating:        number | null;
    color:         string;
    episodeId:     string;
    episodeTitle:  string;
    episodeNumber: number;
    genres:        string[];
    type:          Type;
}

export interface Title {
    romaji:        string;
    english:       null | string;
    native:        string;
    userPreferred: string;
}

export enum Type {
    Ona = "ONA",
    Tv = "TV",
}